Fix underflow in reverse result iterator.
This may have been causing crashes in test82 on gcc 7.2.0.  A reverse
iterator is implemented as a regular iterator, but at one element before
the element it's actually accessing.  (This can really do your head in
because the types' ideas of "forward" and "up" are reversed.)

With a reverse iterator, rbegin() points to end() - 1.  But what if the
container is empty?  The simple, ugly fix is to allow m_index to be -1.
Bear in mind that dereferencing that iterator would be illegal anyway.

+++ include/pqxx/row.hxx
@@ -191,7 +191,11 @@ protected:
   /// Result set of which this is one row.
   result m_result;
   /// Row number.
-  size_t m_index = 0;
+  /**
+   * You'd expect this to be a size_t, but due to the way reverse iterators
+   * related to regular iterators, it must be allowed to underflow to -1.
+   */
+  long m_index = 0;
   /// First column in slice.  This row ignores lower-numbered columns.
   size_type m_begin = 0;
   /// End column in slice.  This row only sees lower-numbered columns.
